diff2 {matrixStats}	R Documentation

Fast lagged differences

Description

Computes the lagged and iterated differences.

Usage

diff2(x, idxs = NULL, lag = 1L, differences = 1L, ..., useNames = TRUE)

Arguments

x
A numeric vector of length N.

idxs
A vector indicating subset of elements to operate over. If NULL, no subsetting is done.

lag
An integer specifying the lag.

differences
An integer specifying the order of difference.

...
Not used.

useNames
If NA, the default behavior of the function about naming support is remained. If FALSE, no naming support is done. Else if TRUE, names attributes of result are set.

Value

Returns a numeric vector of length N - differences.

Author(s)

Henrik Bengtsson

See Also

diff().

Examples

diff2(1:10)

[Package matrixStats version 1.0.0 Index]