matrixStats.options {matrixStats}	R Documentation

Options used for matrixStats

Description

Below are the R options and environment variables that are used by the matrixStats package and packages enhancing it.

WARNING: Note that the names and the default values of these options may change in future versions of the package. Please use with care until further notice.

Options for controlling deprecation

matrixStats.center.onUse:

(string) Action taken when argument center is specified. If "defunct", an error is thrown. If "deprecated", a warning is signaled. If "ignore", it's silently ignored. (Default: "ignore")

matrixStats.center.onScalar:

(string) Action taken when argument center is a scalar. If "defunct", an error is thrown. If "deprecated", a warning is signaled. (Default: "deprecated")

matrixStats.formula.onMistake:

(string) Action taken when argument center is specified with the wrong assumptions of the underlying formula used internally. If "defunct", an error is thrown. If "deprecated", a warning is signaled. (Default: "defunct")

matrixStats.formula.freq:

(numeric) Controls how often the above assumption is checked. (Default: 50 - every 50:th call starting with the first)

matrixStats.ties.method.missing:

(string) Controls whether argument ties.method for colRanks() and rowRanks() should be explicitly specified. If "defunct", an error is produced, if not. If "deprecated", a warning is signaled. If "ignore", it's silently ignored. (Default: "deprecated" in R (>= 4.4.0), otherwise "ignore")

matrixStats.ties.method.freq:

(numeric) Controls how often the above validation is checked. (Default: 10 - every 10:th call starting with the first)

Environment variables that set R options

All of the above R matrixStats.* options can be set by corresponding environment variable R_MATRIXSTATS_* when the matrixStats package is loaded. For example, if R_MATRIXSTATS_TIES_METHOD_FREQ=10, then option matrixStats.ties.method.freq is set to 10 (integer).

Examples

## Not run: 
options(matrixStats.ties.method.freq = 10L)

## End(Not run)

[Package matrixStats version 1.5.0 Index]