rowCumsums {matrixStats}	R Documentation

Cumulative sums, products, minima and maxima for each row (column) in a matrix

Description

Cumulative sums, products, minima and maxima for each row (column) in a matrix.

Usage

rowCumsums(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

colCumsums(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

rowCumprods(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

colCumprods(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

rowCummins(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

colCummins(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

rowCummaxs(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

colCummaxs(x, rows = NULL, cols = NULL, dim. = dim(x), ..., useNames = NA)

Arguments

x
An NxK matrix or, if dim. is specified, an N * K vector.

rows
A vector indicating subset of rows to operate over. If NULL, no subsetting is done.

cols
A vector indicating subset of columns to operate over. If NULL, no subsetting is done.

dim.
An integer vector of length two specifying the dimension of x, also when not a matrix. Comment: The reason for this argument being named with a period at the end is purely technical (we get a run-time error if we try to name it dim).

...
Not used.

useNames
If NA, the default behavior of the function about naming support is remained. If FALSE, no naming support is done. Else if TRUE, names attributes of result are set.

Value

Returns a numeric NxK matrix of the same mode as x, except when x is of mode logical, then the return type is integer.

Author(s)

Henrik Bengtsson

See Also

See cumsum(), cumprod(), cummin(), and cummax().

Examples

x <- matrix(1:12, nrow = 4, ncol = 3)
print(x)

yr <- rowCumsums(x)
print(yr)

yc <- colCumsums(x)
print(yc)

yr <- rowCumprods(x)
print(yr)

yc <- colCumprods(x)
print(yc)

yr <- rowCummaxs(x)
print(yr)

yc <- colCummaxs(x)
print(yc)

yr <- rowCummins(x)
print(yr)

yc <- colCummins(x)
print(yc)

[Package matrixStats version 0.60.0 Index]