signTabulate {matrixStats}	R Documentation

Calculates the number of negative, zero, positive and missing values

Description

Calculates the number of negative, zero, positive and missing values in a numeric vector. For double vectors, the number of negative and positive infinite values are also counted.

Usage

signTabulate(x, idxs = NULL, ...)

Arguments

x
a numeric vector.

idxs
A vector indicating subset of elements to operate over. If NULL, no subsetting is done.

...
Not used.

Value

Returns a named numeric vector.

Author(s)

Henrik Bengtsson

See Also

sign().

[Package matrixStats version 0.60.0 Index]