validateIndices {matrixStats}	R Documentation

Validate indices

Description

Computes validated positive indices from given indices.

Usage

validateIndices(idxs = NULL, maxIdx, allowOutOfBound = TRUE)

Arguments

idxs
A vector indicating subset of elements to operate over. If NULL, no subsetting is done.

maxIdx
The possible max index.

allowOutOfBound
Allow positive out of bound to indicate NA.

Details

WARNING: This function is defunct and will be removed in a future version.

Value

Returns a validated integers list indicating the indices. If some of the indices cannot be represented as an integer, the indices are returned as doubles.

[Package matrixStats version 1.0.0 Index]