rowOrderStats {matrixStats}	R Documentation

Gets an order statistic for each row (column) in a matrix

Description

Gets an order statistic for each row (column) in a matrix.

Usage

rowOrderStats(x, rows = NULL, cols = NULL, which, dim. = dim(x), ...)

colOrderStats(x, rows = NULL, cols = NULL, which, dim. = dim(x), ...)

Arguments

x
A numeric NxK matrix.

rows, cols
A vector indicating subset of rows (and/or columns) to operate over. If NULL, no subsetting is done.

which
An integer index in [1,K] ([1,N]) indicating which order statistic to be returned.

dim.
An integer vector of length two specifying the dimension of x, also when not a matrix.

...
Not used.

Details

The implementation of rowOrderStats() is optimized for both speed and memory. To avoid coercing to doubles (and hence memory allocation), there is a unique implementation for integer matrices.

Value

Returns a numeric vector of length N (K).

Missing values

This method does not handle missing values, that is, the result corresponds to having na.rm = FALSE (if such an argument would be available).

Author(s)

The native implementation of rowOrderStats() was adopted by Henrik Bengtsson from Robert Gentleman's rowQ() in the Biobase package.

See Also

See rowMeans() in colSums().

[Package matrixStats version 0.54.0 Index]