rowMads {matrixStats}	R Documentation

Standard deviation estimates for each row (column) in a matrix

Description

Standard deviation estimates for each row (column) in a matrix.

Usage

rowMads(x, rows = NULL, cols = NULL, center = NULL, constant = 1.4826,
  na.rm = FALSE, dim. = dim(x), ..., useNames = NA)

colMads(x, rows = NULL, cols = NULL, center = NULL, constant = 1.4826,
  na.rm = FALSE, dim. = dim(x), ..., useNames = NA)

rowSds(x, rows = NULL, cols = NULL, na.rm = FALSE, center = NULL,
  dim. = dim(x), ..., useNames = NA)

colSds(x, rows = NULL, cols = NULL, na.rm = FALSE, center = NULL,
  dim. = dim(x), ..., useNames = NA)

Arguments

x
An NxK matrix or, if dim. is specified, an N * K vector.

rows
A vector indicating subset of rows to operate over. If NULL, no subsetting is done.

cols
A vector indicating subset of columns to operate over. If NULL, no subsetting is done.

center
(optional) The center, defaults to the row means for the SD estimators and row medians for the MAD estimators.

constant
A scale factor. See mad for details.

na.rm
If TRUE, missing values are excluded.

dim.
An integer vector of length two specifying the dimension of x, also when not a matrix. Comment: The reason for this argument being named with a period at the end is purely technical (we get a run-time error if we try to name it dim).

...
Additional arguments passed to rowMeans() and rowSums().

useNames
If NA, the default behavior of the function about naming support is remained. If FALSE, no naming support is done. Else if TRUE, names attributes of result are set.

Value

Returns a numeric vector of length N (K).

Author(s)

Henrik Bengtsson

See Also

sd, mad and var. rowIQRs().

[Package matrixStats version 0.60.0 Index]